Added unsigned char 4 to DataBlocks and squelched a warnign in ITexture
Ignore-this: edf5f1eb9997548ed32edc91f1663aea

darcs-hash:20101109105456-ceda8-48f19116d41e75b9942377e2c05ffea5bd53c8c2.gz

+++ src/Resources/ITexture.h
@@ -262,7 +262,7 @@ namespace OpenEngine {
             // Serialization
             virtual void Serialize(IArchiveWriter&) {}; // Todo make pure virtual.
             virtual void Deserialize(IArchiveReader&) {};
-            virtual unsigned int GetSerialzationTag() {return -1;}
+            virtual unsigned int GetSerialzationTag() {return (unsigned int)-1;}
             
             static std::string colorFomatToString(ColorFormat cf) {
                 switch(cf) {
